Refactoring
 * Dropped "cloudbees" from class names 
 * Added Licence headers
 * Updated @since to reflect target release verion
