+++ src/main/java/jenkins/advancedqueue/jobinclusion/strategy/AllJobsJobInclusionStrategy.java
@@ -32,7 +32,7 @@ import org.kohsuke.stapler.DataBoundConstructor;
 
 /**
  * @author Magnus Sandberg
- * @since 2.7
+ * @since 3.0
  */
 public class AllJobsJobInclusionStrategy extends JobInclusionStrategy {
